This is a pretty significant change to the API::Method base class and the derived classes. I will update the wiki tomorrow.
API::Method now inherits from both QObject and QXmlDefaultHandler. This
is the one case where multiple inheritance is supported and actually
encouraged in Qt.

Two virtual methods from QXmlDefaultHandler are implemented in each
API::Method derived class: startElement() and endElement(). These
methods handle pulling the data out of the XML as it's parsed. The way I
set the inheritance up, that's all you have to do in the derived class.

The API::Method class now has its own signal
methodComplete(API::Method*) that it throws once the reply has been
parsed and processed. The caller attaches this to its own slot that
takes an API::Method* argument.

The downside to this is that you have to cast the API::Method* back to
the actual derived class in your slot. I think I'm going to change this
to where the slot is virtual in API::Method and each derived class sends
itself rather than the base class. This will mean you need a slot for
each method type ... but that seems logical anyway. I need to go to
sleep, so I'll do that tomorrow.

The application currently builds and the TestQueryConsole constructor
shows how to use the new API::Method design by calling API::Friends::Get
and displaying the results.