UI work. Ugh.
I've gone about 3 different ways with this, and decided Qt widgets and
Webview where needed is the way I want to go. I had thought of just
using Webkit for everything, but really didn't like the idea of having
to much around with javascript embedded in a web browser embedded in a
C++ application...

You can now pull up all your Notifications in a scrolling, resizing
windwow. I made a mess of a bunch of things getting here, will need to
clean up, but this is a good commit point