Complete revamp of the architecture.
The NotificationCenter will now drive the applicaiton, displaying new
notifications as well as being the jump-off point for navigation,
sending signals to other views when they need to be displayed, etc. I
still have to revamp the notification to the tray icon and then drop the
notification view I had originally built, but it's almost to the point
where we can start rolling on features.

Fixed some other things, came up with some display widgets, made it so
FQL queries will work with the existing API call class structure.