Fixed a couple little problems with the NotificationCenter display, and timing issues getting new notifications.
Added an animated box around new notifications. It's not done yet, you
can't turn it off :) While I was putting off making things pretty, I
wanted to see if I understood the graphics context / painting stuff.
Apparently I do.

Next is to handle marking things as read when the view is activated, and
signaling the tray icon when that happens to reduce the new count.