Yeah. Big changes.
We now directly long into facebook and get the login cookies, as well as
scraping the info we need to be able to connect to the COMET server. 

This actually gets around the f'ing facebook connect bugs at the moment.
There's a bit of a hack involved, and I'm sure it'll break when they fix
their login, but for the moment it's all good.

Next is to change NotificationCheck to use the COMET server and just
listen and spit out stuff.