Made a nice clean interface for our threads.
UTIL::Worker is the base class to derive "runnable" objects from. It's
derived from QObject so it supports signals/slots. It has a single pure
virtual method: void init(). You implement this in your class to do any
setup when the thread starts.

UTIL::WorkerThread is derived from QThread and takes a UTIL::Worker* as
an argument in the constructor and moves the UTIL::Worker into the
QThread. It overrides QThread::run() and calls the UTIL::Worker's init()
method before starting it's event loop.

See CometConnection for an example, and in qtfacebook.cpp for how it's
used.