So ... a few changes.
Finished ripping out the CometConnector stuff. I left the Buddy and
ChatMessage objects, figured we might be able to re-use them. This also
eliminates the dependency on qjson. 

Stream Post Updates (new posts on your facebook page) are working for
basic posts (No external links, media, etc). Viewing Basic Stream posts
and the associated comments is also working. Nothing fancy, but
functional. 

There's a new UTIL::FbUserPicCache singleton class. It's actually pretty
cool. Cuts down on the loadtimes for everything. It's currently used in
the CommentWidget, StreamPostWidget, and in NotificationCenter. The
singleton pattern gives us a nice global cache that can be used anywhere
in the application. 

I also created a new abstract interface for the NotificationCenter
display. I still need to go through and trim out a couple of my original
classes.

Still lots to do but it's starting to resemble a useful application.