We no longer query the comment table in API::FQL::GetStreamPosts. Instead we take the 2-3 "free" ones from the stream table.
Additional comments will be queried upon user request from a
StreamPostWdget, notifications of comments, first views of posts that
have no comments at the time of query, when a user submits a comment,
and re-views of posts that the user has asked for additional comments.
Working on that part now.