Went round and round trying to get the full photo info from the intial multiquery for posts that have photos attached. Survey says: isn't going to happen. multiquery doesn't handle quesies against other queries where the colum in question isn't always there (in this case, the photo object in the attachement).
The StreamPostWidget will now emit the PhotoId, AlbumId, and OwnerId
back to the NotificationCenter when a user clicks on a photo in a post.
This decouples the (to be written) PhotoView and AlbumView widgets from
the StreamPost Widget, which I think is the the right way to do it.
Those widgets will have to do an FQL query at startup for all the photo
and album info. 

There's still a lot to do in the StreamPostWidget but I'm about sick of
working on that object at the moment :) I think I'm going to do the
PhotoViewWidget and see if I've learned anything so far :)