Image viewing ... is getting there.
The real problem is that half the photo id's returned by facebook
currently don't work. It seems they're in the middle of converting to
their new id system, and the whole thing is sort of horked.

If the pid (photo id) is in the new <userId>_<objectId> format ... it
works and you can query the photo table for the info. If it's a 64 bit
int (the old format) ... it usually won't return anything from the photo
table (but sometimes it does ...). ::sigh::