Improved the FbPhotoViewWidget quite a bit. I think I'm actually going to move the FQL query into it and pass it the photo id (pid) in the constructor. Unsure yet. I can't decide if the widget should "own" the photo data or not. I'm almost thinking of modifying the FbPhotoCache to cache the whole FbPhoto object as well as the images (mainly for album viewing) but I don't know if there's real value in that or not. Even caching the images is a fairly narrow use case, but it's cheap to do.
I wrote a singleton (UTIL::OurUserInfo) to replace the UserInfo object,
but haven't gone through and converted everything yet. We're passing the
UserInfo pointer to everything under the sun mainly to give to
API::Factory objects, so it makes more sense to just make it a singleton
and let the factory have direct access to it.